Fixed exception when removing control from views
The willMoveToSuperview method used to layout the control when it is
added to its superview is also called with nil when being removed from
the superview. This sometimes lead to exception in the call
moveThumbToIndex ad the end of the method.
This fix merely returns from willMove when the newSuperView is nil.